Adds javascript function. Incomplete.
javascript is incomplete. Only completed creation
of the board, creation of players, winning condition,
assignment of players.

Current Issue: TOKEN IS NOT READING BECAUSE THE INSTANTIATION
OF PLAYERS INPUT IS LATE. WE PLAYROUND BEFORE WE START
PLAYERSINPUT. MAYBE WE NEED TO INSTANTIATE PLAYERSINPUT
IN PLAYROUND. IDK. THEY SUGGEST CREATE A START BUTTON TO FIX.
RECONSIDER TMRO.

Initially, this was working fine but it broke after I created
factory function to create players and allow player to assign
name and token.

Started this project on: 8 July.
D1: CSS + html
D2: ONE function
D3: ONE function
D4: did PlayerInput and WinningConditions
but broke the GameController().