Add JS functions up to Step 6/7
I fixed the player input to take in data upon
submit button click. When click is heard, it pushes
on to players array the data retrieved from form.

**Important: need to preventdefault() when using
the submit button.

I let players to be accessible by gamecontroller.
In gamecontroller, I added another submit button
event listener. I push the info from PlayersInput()
into the players array by using

const temp = Object.entries(playersInput);
(this is convert object back to array)
Then I pushed it on the array by going deep into
the array temp[0][1][0].
first part = array when retrieved into gamecontroller
second part is the array in playerinput before exported
third part is the actual data of the first object in the
array.

This ^ part was really hard to figure out.

I also assigned the activeplayer in the submit button
so that it is reassigned for when I run playRound()
which was also in the submit button.

TLDR: submit button catch all the stuff that was not
instantiated before it was created.

In winningconditions, i added a new block of else if
for DRAW. all blocks must have text content.

In gameend I just window.location.reload(). A really
cheap way to reset the whole program.

!+ TO ADD:
1.I should try to figure out how to reset everything
and save the player name...but I rather work on the
AI player.

2. AI player

3. Smart AI player