Add NPC as playerTwo Created a npcChoice function where it randoms btwn 0 and 9 (same size as board array) and then fills out the board, block and board array with data from the npc choice.
if 0 to 9 is not empty, we hop out of the RECURSIVE
loop (AHAHAHA, I USED RECURSION. MUCH PROUD.)

else we go back to random a number again.

In game controller i encounter a pblm where it would
always say that the winner is the NPC and not me.
So, i add an if statement to check if i have already
won, and if i haven't win (the modal is not active),
then it will run npcChoice function check if they won,
switch players, print the round. then back to me.

TLDR: Almost all done already. Last step is the AI.

!+ I would prefer if the npc took longer to write
its result. It's instant right now.